Refactor: Separate views into individual files
- Refactor: update content of views
