+++ iosApp/iosApp/ContentView.swift
@@ -3,7 +3,7 @@ import SwiftUI
 
 struct ContentView: View {
     
-    // MARK: PROPERTIES
+// MARK: PROPERTIES
     @State var selectedTab: Int = 0
 //	let greet = Greeting().greet()
 
@@ -11,7 +11,7 @@ struct ContentView: View {
         case firstTab = 0, secondTab, thirdTab, fourthTab
     }
     
-    // MARK: BODY
+// MARK: BODY
 	var body: some View {
 
         TabView(selection: $selectedTab) {
@@ -23,16 +23,14 @@ struct ContentView: View {
                 }
                 .tag(Tab.firstTab.rawValue)
             
-            LearningView()
-                .font(.largeTitle)
-                .foregroundColor(.blue)
+            LearnView(selectedTab: $selectedTab)
                 .tabItem {
                     Image(systemName: "graduationcap.fill")
                     Text("Learn")
                 }
                 .tag(Tab.secondTab.rawValue)
             
-            MenuView()
+            MenuView(selectedTab: $selectedTab)
                 .font(.largeTitle)
                 .foregroundColor(.blue)
                 .tabItem{
@@ -40,108 +38,8 @@ struct ContentView: View {
                     Text("Menu")
                 }
                 .tag(Tab.thirdTab.rawValue)
-            
         }
-        
-        
 	}
-}
-
-
-struct WalletView: View {
-    
-    @EnvironmentObject var viewModel: WalletViewModel
-    @Binding var selectedTab: Int
-    
-    var body: some View {
-        ZStack {
-            //Color.red.ignoresSafeArea()
-            VStack {
-                
-                Image("Logo")
-                Spacer()
-                
-                
-                Text("Block Height")
-                    .font(.largeTitle)
-                    .foregroundColor(.blue)
-                
-                Text("\(String(viewModel.blockHeight))")
-                    //.font(.largeTitle)
-                    .font(.system(size: 60))
-                    .foregroundColor(.blue)
-                
-
-                Spacer()
-                
-            }
-        }
-        .onAppear(perform: viewModel.load)
-    }
-}
-struct LearningView: View {
-    
-    @State var isSelected: Bool = false
-    
-    var body: some View {
-        
-        VStack(spacing: 40) {
-            
-            Text("Learning Tab")
-            
-            Spacer()
-        }
-        .padding(40)
-        
-    }
-}
-
-struct MenuView: View {
-    
-    @State var isSelected: Bool = false
-    
-    var body: some View {
-        
-        VStack(spacing: 40) {
-            
-            RoundedRectangle(cornerRadius: 25)
-                .frame(height: 200)
-                .foregroundColor(isSelected ? Color.green : Color.red)
-            
-            Button(action: {
-                isSelected.toggle()
-            }, label: {
-                Text("Button")
-                    .font(.headline)
-                    .foregroundColor(.white)
-                    .frame(height: 55)
-                    .frame(maxWidth: .infinity)
-                    .background(Color.blue)
-                    .cornerRadius(25)
-            })
-            
-            Text("Double TAP Gesture")
-                .font(.headline)
-                .foregroundColor(.white)
-                .frame(height: 55)
-                .frame(maxWidth: .infinity)
-                .background(Color.blue)
-                .cornerRadius(25)
-//                .onTapGesture {
-//                    isSelected.toggle()
-//                }
-                .onTapGesture(count: 2, perform: {
-                    isSelected.toggle()
-                })
-            
-            Text("Menu Item 1")
-            Text("Menu Item 2")
-            
-            Spacer()
-        }
-        .padding(40)
-        
-    }
 }
 
 // MARK: PREVIEW
