+++ iosApp/iosApp/MenuView.swift
@@ -1,0 +1,26 @@
+//
+//  MenuView.swift
+//  iosApp
+//
+//  Created by geigerzaehler on 2023-12-07.
+//  Copyright © 2023 orgName. All rights reserved.
+//
+
+import SwiftUI
+
+struct MenuView: View {
+    
+    @EnvironmentObject var viewModel: WalletViewModel
+    @Binding var selectedTab: Int
+    
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+            .font(.largeTitle)
+            .foregroundColor(.blue)
+    }
+}
+
+#Preview {
+    MenuView(selectedTab: .constant (0))
+        .environmentObject(WalletViewModel())
+}
