+++ iosApp/iosApp/WalletView.swift
@@ -1,0 +1,63 @@
+//
+//  WalletView.swift
+//  iosApp
+//
+//  Created by geigerzaehler on 2023-12-07.
+//  Copyright © 2023 orgName. All rights reserved.
+//
+
+import SwiftUI
+
+struct WalletView: View {
+    
+    @EnvironmentObject var viewModel: WalletViewModel
+    @Binding var selectedTab: Int
+    
+    @State var isSelected: Bool = false
+    
+    var body: some View {
+        VStack(spacing: 40) {
+            
+            RoundedRectangle(cornerRadius: 25)
+                .frame(height: 200)
+                .foregroundColor(isSelected ? Color.green : Color.red)
+            
+            Button(action: {
+                isSelected.toggle()
+            }, label: {
+                Text("Button")
+                    .font(.headline)
+                    .foregroundColor(.white)
+                    .frame(height: 55)
+                    .frame(maxWidth: .infinity)
+                    .background(Color.blue)
+                    .cornerRadius(25)
+            })
+            
+            Text("Double TAP Gesture")
+                .font(.headline)
+                .foregroundColor(.white)
+                .frame(height: 55)
+                .frame(maxWidth: .infinity)
+                .background(Color.blue)
+                .cornerRadius(25)
+//                .onTapGesture {
+//                    isSelected.toggle()
+//                }
+                .onTapGesture(count: 2, perform: {
+                    isSelected.toggle()
+                })
+            
+            Text("Wallet Item 1")
+            Text("Wallet Item 2")
+            
+            Spacer()
+        }
+        .padding(40)
+    }
+}
+
+#Preview {
+    WalletView(selectedTab: .constant (0))
+        .environmentObject(WalletViewModel())
+}
